+++ CHANGELOG.md
@@ -14,6 +14,7 @@ All notable changes to this project will be documented in this file.
 - [GUI] Extend -single-instance fix to 1.10 (#3153 by: DasSkelett; reviewed: HebaruSan)
 - [GUI] Escape '&' in mod list, fix path in GUIConfig.xml exception (#3149 by: DasSkelett; reviewed: HebaruSan)
 - [GUI] Print message for unknown install exception (#3164 by: DasSkelett; reviewed: HebaruSan)
+- [GUI] Search timer for Linux (#3167 by: HebaruSan; reviewed: DasSkelett)
 
 ### Internal